🚚 コードを外部に移動
